Move LilyPond version manager out of the preferences
The reasoning is that most users need to use this at some point, unlike
the more advanced options in the preferences.
